LayoutTests/fast/profiler/multiple-and-different-scoped-anonymous-function-calls-expected.txt

This page's JavaScript calls the same anonymous function from multiple and differently scoped locations. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be a call to fakeObject with fakeInteriorFunction as its child, but call() should not be shown.

Profile title: Multiple and different scoped calls to the same anonymous function
Thread_1 (no file) (line 0)
   startTest multiple-and-different-scoped-anonymous-function-calls.html (line 11)
      anonymousFunction profiler-test-JS-resources.js (line 29)
         insertNewText profiler-test-JS-resources.js (line 17)
            createElement (no file) (line 0)
            createTextNode (no file) (line 0)
            appendChild (no file) (line 0)
            getElementById (no file) (line 0)
      intermediaryFunction profiler-test-JS-resources.js (line 33)
         anonymousFunction profiler-test-JS-resources.js (line 29)
            insertNewText profiler-test-JS-resources.js (line 17)
               createElement (no file) (line 0)
               createTextNode (no file) (line 0)
               appendChild (no file) (line 0)
               getElementById (no file) (line 0)
      endTest profiler-test-JS-resources.js (line 1)
